requirements.txt
Flask==3.1.2
flask-login==0.6.3
flask-wtf==1.2.1
mitmproxy==12.1.2
bcrypt==4.2.1
protobuf==5.29.3
cryptography==45.0.6
pyOpenSSL==25.1.0
requests==2.32.3
gunicorn==23.0.0